Module fadroma_declare_contract::scrt_contract_api

Macros that help with defining the contract’s API.

message! and messages!

The message! and messages! macros depend on derive macros from serde and schemars. This complicates the reexport via fadroma core due to how procedural macros resolve crates (see https://github.com/serde-rs/serde/issues/1465#issuecomment-458420543)

The workaround at https://github.com/serde-rs/serde/issues/1465#issuecomment-800686252 is automatically implemented in the mod msg generated by the contract! macro This is done automatically in the mod msg defined by the contract! macro. However if you use the message macros manually, it is up to you to implement a corresponding workaround.

define_*_message(s)

These macros are called internally by the contract! macro and get passed the necessary values from the corresponding contract! sections (which define API and implementation in the same place) to define just the API.

Alternatively if you have your message definitions in a separate module, (e.g. if you have a separate API crate defining the interface for multiple contracts), these macros allow for contract! to just import and use the external message structs/enums.